Add ability to run app as standalone proxy
Scripted through Gulp, so it can both watch for changes and proxy to AM so that the app does not actually need to be deployed on the server.
Also splits off the vendor-supplied libraries (jQuery and crypto-js) into their own chunk so that it doesn't bloat the JS for each sample.